+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### UNRELEASED
 - Add dependant processes. See [docs](https://docs.rs/steward/latest/steward/dep/index.html).
+- Add [`Cmd::spawn`](https://docs.rs/steward/latest/steward/cmd/struct.Cmd.html#method.spawn) and [`Process::spawn`](https://docs.rs/steward/latest/steward/process/struct.Process.html#method.spawn) methods.
+- Expose [`RunningProcess`](https://docs.rs/steward/latest/steward/process/struct.RunningProcess.html). Add [`RunningProcess::stop`](https://docs.rs/steward/latest/steward/process/struct.RunningProcess.html#method.stop) (`unix` only, for now).
 
 ### 0.0.3
 - Improve process pool output.